Replace static domain config with DB-backed lookup; add Domains UI
Alias creation now accepts domain_id (ULID) resolved from the
CustomDomains repository instead of a domain name from a static
config list. Removes Domains []string from APIConfig and MilterConfig.

- Add GetByName to CustomDomainsReader; implement in GORM and cached
  drivers
- Add duplicate-name guard and auto-verify on domain create
- Add global/verified filter params to CustomDomainFilter
- Rename OpenAPI domainSource→domainType, domain→domain_id in
  createAliasRequest; regenerate model.gen.go
- Add DomainsTab and AddDomainForm Vue components wired into MainTabs
- Fix AddAliasForm: import missing Dropdown component, switch Domain
  selector to CFormSelect backed by domain ULID
- Add API error modal to AliasesTab, ApiKeysTab, PrAddrsTab, UsersTab
- Add socketmap subcommand skeleton with SocketMapConfig

+++ internal/applications/rest/model_utils_test.go
@@ -73,17 +73,17 @@ func TestPgmTMetadata(t *testing.T) {
 		TotalRecords: 42,
 	}
 	result := pgmTMetadata(pgm)
-	assert.Equal(t, float32(2), result.CurrentPage)
-	assert.Equal(t, float32(1), result.FirstPage)
-	assert.Equal(t, float32(5), result.LastPage)
-	assert.Equal(t, float32(10), result.PageSize)
-	assert.Equal(t, float32(42), result.TotalRecords)
+	assert.Equal(t, 2, result.CurrentPage)
+	assert.Equal(t, 1, result.FirstPage)
+	assert.Equal(t, 5, result.LastPage)
+	assert.Equal(t, 10, result.PageSize)
+	assert.Equal(t, 42, result.TotalRecords)
 }
 
 func TestPgmTMetadata_Zero(t *testing.T) {
 	result := pgmTMetadata(entities.PaginationMetadata{})
-	assert.Equal(t, float32(0), result.CurrentPage)
-	assert.Equal(t, float32(0), result.TotalRecords)
+	assert.Equal(t, 0, result.CurrentPage)
+	assert.Equal(t, 0, result.TotalRecords)
 }
 
 func TestUserTResponse(t *testing.T) {
